rename stats sliver comp file
