+++ src/client/components/HomePage/Statistics.jsx
