+++ src/client/components/HomePage/index.jsx
@@ -8,7 +8,7 @@ import homeActions from '~/actions/home'
 import loginActions from '~/actions/login'
 import { USER_PAGE } from '~/util/types'
 import TrustedBySliver from './TrustedBySliver'
-import FaqSliver from './FaqSliver'
+import FaqSliver from './Statistics'
 import DescriptionSliver from './FeatureListSliver'
 import Section from '../Section'
 import LandingGraphicSliver from './LandingGraphicSliver'
